// src/schemas/vendor/vendorMaster.schema.ts
import * as z from "zod"

export const VendorMasterFormSchema = z.object({
  cmp_code: z.string().min(1, "Company Code is required"),
  cmp_name: z.string().min(1, "Company Name is required"),
  vendor_code: z.string().min(1, "Vendor Code is required"),
  name: z.string().min(1, "Name is required"),

  search_name: z.string().nullable().optional(),
  address: z.string().nullable().optional(),
  address_2: z.string().nullable().optional(),
  city: z.string().nullable().optional(),
  contact: z.string().nullable().optional(),
  phone_no: z.string().nullable().optional(),
  currency_code: z.string().nullable().optional(),
  language_code: z.string().nullable().optional(),
  registration_no: z.string().nullable().optional(),
  payment_terms_code: z.string().nullable().optional(),
  country_region_code: z.string().nullable().optional(),
  blocked: z.string().nullable().optional(),
  pay_to_vendor_no: z.string().nullable().optional(),
  payment_method_code: z.string().nullable().optional(),
  last_date_modified: z.string().nullable().optional(),
  vat_registration_no: z.string().nullable().optional(),
  post_code: z.coerce.number().nullable().optional(),
  county: z.string().nullable().optional(),
  eori_number: z.string().nullable().optional(),
  email: z.string().email("Invalid email").nullable().optional(),
  home_page: z.string().url("Invalid URL").nullable().optional(),
  primary_contact_no: z.string().nullable().optional(),
  mobile_phone_no: z.string().nullable().optional(),
  location_code: z.string().nullable().optional(),
  gst_registration_no: z.string().nullable().optional(),
  gst_vendor_type: z.string().nullable().optional(),
  aggregate_turnover: z.string().nullable().optional(),
  arn_no: z.string().nullable().optional(),
  transporter: z.boolean().nullable().optional(),
  subcontractor: z.boolean().nullable().optional(),
  assessee_code: z.string().nullable().optional(),
  pan_no: z.string().nullable().optional(),
  pan_status: z.string().nullable().optional(),
  pan_reference_no: z.string().nullable().optional(),
  state_code: z.string().nullable().optional(),
  tax_code: z.string().nullable().optional(),
  fssi_code: z.string().nullable().optional(),
  fssi_expiry_date: z.string().nullable().optional(),
  msme_no: z.string().nullable().optional(),
  msme_status: z.string().nullable().optional(),
  msme_type: z.string().nullable().optional(),
  msme_applicable: z.boolean().nullable().optional(),
  email_2: z.string().email("Invalid secondary email").nullable().optional(),
  api_gst_reg_no: z.string().nullable().optional(),
  status: z.string().nullable().optional(),
})

export type VendorMasterFormValues = z.infer<typeof VendorMasterFormSchema>;